rpc: move MarshalAsset function to new file taprpc/marshal.go
This refactor allows us to avoid an import cycle when the RPC proof
courier marshals an asset.
